% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sparseAHC.R
\name{isSymmetric.dgCMatrix}
\alias{isSymmetric.dgCMatrix}
\title{Test if a Matrix of type dgCMatrix is Symmetric
Test if a Matrix of type dgCMatrix is Symmetric}
\usage{
\method{isSymmetric}{dgCMatrix}(object, tol = 100 * .Machine$double.eps, ...)
}
\arguments{
\item{object}{any R object; a ‘matrix’ for the matrix method.}

\item{tol}{numeric scalar >= 0.  Smaller differences are not considered.}

\item{...}{further arguments passed to methods; the matrix method passes.}
}
\description{
Test if a Matrix of type dgCMatrix is Symmetric
Test if a Matrix of type dgCMatrix is Symmetric
}

