% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sparseHC}
\alias{sparseHC}
\title{agglomerative hierarchical clustering for sparse similarity matrices}
\usage{
sparseHC(S, linkage = c("average", "single", "complete"), noOrder = FALSE)
}
\arguments{
\item{S}{sparse similarity matrix}

\item{linkage}{is c(average, single, complete)}

\item{noOrder}{if TRUE does not retun the order of points which is required for plotting dendrogram to save memory and time}
}
\value{
hclust object
}
\description{
agglomerative hierarchical clustering for sparse similarity matrices
}
\examples{
library("Matrix");
library("igraph");
library("sparseHC");

A = Matrix(0, nrow = 6, ncol = 6, sparse = TRUE);
A[1,2]= 2; A[1,5]= 3; A[1,4]= 2.5; A[2,5]= 4;
A[3,4]= 2; A[3,5]= 6; A[4,5]= 5; A[5,6]= 4.6;
A = A + t(A);

G = graph.adjacency(A, mode = "undirected", weighted=TRUE);
plot(G,edge.label=E(G)$weight,vertex.label=V(G)-1)

H=sparseHC(A,"average",FALSE)
plot(H);
}

