% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sparseAHC.R
\name{isSymmetric.dgCMatrix}
\alias{isSymmetric.dgCMatrix}
\title{Test if a matrix of type dgCMatrix is symmetric.}
\usage{
\method{isSymmetric}{dgCMatrix}(object, tol = 100 * .Machine$double.eps, ...)
}
\arguments{
\item{object}{a matrix of type dgCMatrix.}

\item{tol}{numeric scalar >= 0.  Smaller differences are not considered.}

\item{...}{further arguments passed to methods; the matrix method passes.}
}
\description{
Test if a matrix of type dgCMatrix is symmetric.
}

